fix: Resolve uv.lock conflicting dependencies with semgrep (#1234)
* fix: Resolve uv.lock conflicting dependencies with semgrep

This commit addresses dependency resolution conflicts between semgrep and OpenTelemetry packages by:

1. **Makefile improvements**:
   - Updated venv target to check if uv is already installed system-wide
   - Modified test target to use `uv run pytest` instead of installing pytest separately
   - Fixed semgrep target to use `uvx semgrep` to avoid dependency conflicts
   - Removed syntax error in lint-smart target (extra `fi` statement)

2. **pyproject.toml restructuring**:
   - Moved development dependencies from [project.optional-dependencies] to [dependency-groups]
   - Removed semgrep from dev dependencies (now run via uvx to avoid conflicts)
   - Simplified observability extras to only include core OpenTelemetry packages
   - Removed observability-jaeger, observability-zipkin, and observability-all extras
   - Removed dev-all extras group

3. **Documentation updates** (docs/docs/manage/observability/observability.md):
   - Expanded backend options with detailed descriptions of Arize Phoenix, Jaeger, and Grafana Tempo
   - Added clear guidance on exporter backend installation per platform
   - Clarified that only one exporter backend should be installed at a time
   - Improved formatting and structure for better readability

4. **uv.lock regeneration**:
   - Updated lock file to reflect new dependency structure
   - Resolved conflicts between semgrep and opentelemetry-sdk packages

The changes transition towards using uv/uvx for dependency management while maintaining backward compatibility with existing workflows.


* fix: Migrate pytest workflow to uv and resolve type hints

- Update .github/workflows/pytest.yml to use uv instead of pip
- Fix Makefile venv target shell syntax for uv installation
- Update type hints for keep_alive parameter (int -> float)
- Improve test_translate_echo.py with proper type annotations
- Add hypothesis to dev dependencies for property-based testing
- Use unused_tcp_port fixture to prevent port conflicts

Resolves conflicting dependencies in uv.lock and improves type safety.


* fix: Use || true instead of ||:


---------

+++ mcpgateway/translate.py
@@ -623,7 +623,7 @@ class SSEEvent:
 def _build_fastapi(
     pubsub: _PubSub,
     stdio: StdIOEndpoint,
-    keep_alive: int = KEEP_ALIVE_INTERVAL,
+    keep_alive: float = KEEP_ALIVE_INTERVAL,
     sse_path: str = "/sse",
     message_path: str = "/message",
     cors_origins: Optional[List[str]] = None,
@@ -1026,7 +1026,7 @@ async def _run_stdio_to_sse(
     host: str = "127.0.0.1",
     sse_path: str = "/sse",
     message_path: str = "/message",
-    keep_alive: int = KEEP_ALIVE_INTERVAL,
+    keep_alive: float = KEEP_ALIVE_INTERVAL,
     header_mappings: Optional[Dict[str, str]] = None,
 ) -> None:
     """Run stdio to SSE bridge.
@@ -1771,7 +1771,7 @@ async def _run_multi_protocol_server(  # pylint: disable=too-many-positional-arg
     expose_streamable_http: bool = False,
     sse_path: str = "/sse",
     message_path: str = "/message",
-    keep_alive: int = KEEP_ALIVE_INTERVAL,
+    keep_alive: float = KEEP_ALIVE_INTERVAL,
     stateless: bool = False,
     json_response: bool = False,
     header_mappings: Optional[Dict[str, str]] = None,
@@ -2250,7 +2250,7 @@ def start_streamable_http_client(url: str, bearer_token: Optional[str] = None, t
 
 
 def start_stdio(
-    cmd: str, port: int, log_level: str, cors: Optional[List[str]], host: str = "127.0.0.1", sse_path: str = "/sse", message_path: str = "/message", keep_alive: int = KEEP_ALIVE_INTERVAL
+    cmd: str, port: int, log_level: str, cors: Optional[List[str]], host: str = "127.0.0.1", sse_path: str = "/sse", message_path: str = "/message", keep_alive: float = KEEP_ALIVE_INTERVAL
 ) -> None:
     """Start stdio bridge.
